move waiting before upload into createFile function
